Upload materials texture during extraction

+++ src/particle_attributes.rs
@@ -53,5 +53,5 @@ macro_rules! define_attributes_and_swap {
 
 define_attributes_and_swap! {
     velocity: IVec2,
-    momentum: u8
+    momentum: u16
 }
